% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverChange.R
\name{coverChange}
\alias{coverChange}
\title{Gain and loss in land cover classification}
\usage{
coverChange(x, y)
}
\arguments{
\item{x}{First classified image.}

\item{y}{Second classified image of a time later.}
}
\description{
This algorithm is able to obtain gain and loss in land cover classification.
}
\section{Value}{

A list with loss raster, gain raster, loss hectares and gain hectares.
}

\section{Note}{

It is important to mention that the images used to obtain gain and loss in land cover classification
must have the same extent. It is preferable that the classified images come from the same software.
}

\examples{
library(ForesToolboxRS)
library(raster)

rastClas1 <- raster(ncol=10, nrow=10, xmn=-1000, xmx=1000, ymn=-100, ymx=900)
values(rastClas1) <- floor(runif(100, min = 1, max = 5))

rastClas2 <- raster(ncol=10, nrow=10, xmn=-1000, xmx=1000, ymn=-100, ymx=900)
values(rastClas2) <- floor(runif(100, min = 1, max = 5))

change <- coverChange(rastClas1, rastClas2)

}
\author{
Yonatan Tarazona
}
