% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endm}
\alias{endm}
\title{Signatures data}
\format{
A sf object with 60 features and one variable:
\describe{
\item{class}{a value between 1 and 4}
}
}
\usage{
endm
}
\description{
A dataset containing signatures such as forest (cod1), urban (cod4), agriculture (cod3) and water (cod2).
}
\keyword{datasets}
