% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForesToolboxRS-package.R
\docType{package}
\name{ForesToolboxRS-package}
\alias{ForesToolboxRS}
\alias{ForesToolboxRS-package}
\title{ForesToolboxRS: Remote Sensing Tools for Forest Monitoring}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Remote sensing tools for forest monitoring includes a series of 
    novel techniques to detect disturbances in the forest through optical images 
    and fusion of optical and microwave data. All tools are based on scientific 
    publications.
}
\author{
\strong{Maintainer}: Yonatan Tarazona Coronel \email{geoyons@gmail.com} (\href{https://orcid.org/0000-0002-5208-1004}{ORCID})

Other contributors:
\itemize{
  \item Cesar Aybar \email{csaybar@gmail.com} (\href{https://orcid.org/0000-0003-2745-9535}{ORCID}) [contributor]
}

}
\keyword{internal}
