% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionRS.R
\name{fusionRS}
\alias{fusionRS}
\title{Fusion of images with different observation geometry.}
\usage{
fusionRS(x, na = TRUE)
}
\arguments{
\item{x}{RasterStack, RasterBrick}

\item{na}{If TRUE the NA values of the images will be omitted from the analysis.}
}
\description{
This algorithm allows to fusion images coming from different spectral sensors
(e.g., optical-optical, optical and SAR or SAR-SAR). Among many of the qualities
of this function, it is possible to obtain the contribution (%) of each variable
in the fused image.
}
\section{References}{

Tarazona et al...
}

\section{Note}{

Before executing the function, it is recommended that images coming from different
sensors or from the same sensor have a co-registration.
}

\examples{
library(ForesToolboxRS)
library(raster)
rasterio = list(nXOff = 50, nYOff = 50)
img <- system.file("simple_mosaic", package="ForesToolboxRS") \%>\%
  list.files(full.names = TRUE) \%>\%
  read_stars(RasterIO = rasterio)
img <- smootH(img)
fusion <- fusionRS(x=img)
plotRGB(fusion[[1]], axes=FALSE, stretch="lin",main ="Fused images")
}
