% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_optical}
\alias{img_optical}
\title{A landsat 8 image II}
\format{
A RasterBrick object with 6 layers representing optical bands of Landsat 8
}
\usage{
img_optical
}
\description{
A dataset containing a Landsat 8 image with the following bands: blue, green, red, nir, swir1 and swir2
}
\keyword{datasets}
