% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_radar}
\alias{img_radar}
\title{A Sentinel-1 image}
\format{
A RasterBrick object with 4 layers representing bands of Sentinel-1
}
\usage{
img_radar
}
\description{
A dataset containing a radar image (Sentinel-1)
}
\keyword{datasets}
