% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearTrend.R
\name{linearTrend}
\alias{linearTrend}
\title{Linear Trend in Remote Sensing}
\usage{
linearTrend(x, type = "lm", ...)
}
\arguments{
\item{x}{RasterStack, RasterBrick or Matrix (row*col, n_images).}

\item{type}{There are two options: "lm" for linear regressions and "glm" for Fitting
Generalized Linear Models.}

\item{...}{For "lm" and "glm": arguments to be used to form the default control argument
if it is not supplied directly. See \link[stats]{lm} and \link[stats]{glm}.}
}
\description{
Linear trend is useful for mapping forest degradation, land degradation, among others.
This algorithm is capable of obtaining the slope of an ordinary least-squares
linear regression and its reliability (p-value).
}
\section{Details}{

Linear regression is widely used to analyze forest degradation or land degradation.
Specifically, the slope and its reliability are parameters for this analysis and these
can be obtained whit this function. On the other hand, logistic regression allows
obtaining a degradation risk map, in other words, it is a probability map. Please see
the references or see [Tarazona and Maria-Miyasiro (2020)](https://doi.org/10.1016/j.rsase.2020.100337).
}

\section{References}{

Tarazona, Y., Maria, Miyasiro-Lopez. (2020). Monitoring tropical forest degradation using
remote sensing. Challenges and opportunities in the Madre de Dios region, Peru. Remote
Sensing Applications: Society and Environment, 19, 100337.

Wilkinson, G.N., Rogers, C.E., 1973. Symbolic descriptions of factorial models for
analysis of variance. Appl. Stat. 22, 392-399.

Chambers, J.M., 1992. Statistical Models in S. CRS Press.
}

\examples{
library(ForesToolboxRS)
library(raster)

e <- extent(350420.9,352028.8,-1417869,-1416288)
imgs <- crop(img_pv, e)
trend <- linearTrend(x = imgs)
plot(trend[[1]]) # raster of slope
plot(trend[[2]]) # raster of p-value

}
\author{
Yonatan Tarazona
}
