% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla.R
\name{mla}
\alias{mla}
\title{Supervised and unsupervised classification in Remote Sensing}
\usage{
mla(img, endm, model = "svm", training_split = 80, verbose = FALSE, ...)
}
\arguments{
\item{img}{RasterStack or RasterBrick.}

\item{endm}{SpatialPointsDataFrame or SpatialPolygonsDataFrame (typically shapefile)
containing the training data.}

\item{model}{Model to use. It can be \link[e1071]{svm} like \code{model = 'svm'}, \link[randomForest]{randomForest} like
\code{model = 'randomForest'}, \link[e1071]{naiveBayes} like \code{model = 'naiveBayes'}, \link[rpart]{rpart} like
\code{model = 'rpart'}, \link[nnet]{nnet} like \code{model = 'nnet'}, \link[kknn]{train.kknn} like
\code{model = 'train.kknn'}.}

\item{training_split}{For splitting samples into two subsets, i.e. training data and for
testing data.}

\item{verbose}{This paramater is Logical. It Prints progress messages during execution.}
}
\description{
This developed function allows to execute supervised and unsupervised classification
in satellite images through various algorithms.
}
\section{References}{

Gareth James, Daniela Witten, Trevor Hastie, Robert Tibshirani. (2013).
An introduction to statistical learning : with applications in R. New York: Springer.

Mountrakis, G., Im, J., Ogole, C. (2011). Support vector machines in remote sensing:
A review. ISPRS Journal of Photogrammetry and Remote Sensing, 66, 247-259.

Belgiu, M., Dragut., L. (2016). Random Forest in Remote Sensing: A Review of Applications
and Future Directions. ISPRS Journal of Photogrammetry and Remote Sensing, 114, 24-31.

Maxwell, A.E., Warner, T.A., Fang, F. (2018). Implementation of machine-learning
classification in remote sensing: an applied review. International Journal of Remote
Sensing, 29(9), 2784-2817.

Pradhan, R., Ghose, M.K., Jeyaram, A. (2010). Land Cover Classification of Remotely
Sensed Satellite Data using Bayesian and Hybrid classifier. International Journal
of Computer Applications, 7(11).

Holloway, J., Mengersen, K. (2018). Statistical Machine Learning Methods and Remote
Sensing for Sustainable Development Goals: A Review. Remote Sensing, 10(9), 1365.
}

\section{Note}{

}

\examples{
library(ForesToolboxRS)
library(raster)
library(snow)

# Load the dataset
data(FTdata)

# Signatures
soil<-c(8980,8508,8704,13636,16579,11420)
forest<-c(8207,7545,6548,16463,9725,6673)
water<-c(9276,9570,10089,6743,5220,5143)
val <- matrix(c(soil,forest,water), 3, 6, byrow = TRUE)
colnames(val) <- c("B1", "B2", "B3","B4","B5","B6")
endm <- data.frame(val, class = c("soil","forest","water"))

reptime <- 50
indx <- rep(1:nrow(endm), reptime)
endm <- endm[indx, ]

# K-Nearest Neighbor Classifier
knn <- mla(img = img, endm, model = "kknn", kmax = 10, training_split = 80)

}
\author{
Yonatan Tarazona
}
