% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MosaicFreeCloud.R
\name{mosaic-free-cloud}
\alias{mosaic-free-cloud}
\alias{MosaicFreeCloud}
\alias{MosaicFreeCloud.character}
\alias{MosaicFreeCloud.RasterStack}
\alias{MosaicFreeCloud.RasterBrick}
\alias{MosaicFreeCloud.stars}
\title{Get the best Available Pixel on an Image Collection}
\usage{
MosaicFreeCloud(img, time, CLUSTER = NULL, fit_negative = TRUE,
  RasterLayer = TRUE, ...)

\method{MosaicFreeCloud}{character}(img, time, CLUSTER = NULL,
  fit_negative = TRUE, RasterLayer = TRUE, ...)

\method{MosaicFreeCloud}{RasterStack}(img, time, CLUSTER = NULL,
  fit_negative = TRUE, RasterLayer = TRUE, ...)

\method{MosaicFreeCloud}{RasterBrick}(img, time, CLUSTER = NULL,
  fit_negative = TRUE, RasterLayer = TRUE, ...)

\method{MosaicFreeCloud}{stars}(img, time, CLUSTER = NULL,
  fit_negative = TRUE, RasterLayer = TRUE, ...)
}
\arguments{
\item{img}{Character vector, RasterStack, RasterBrick or 3D/4D star object.}

\item{time}{Date or character vector with the dates of the files.}

\item{CLUSTER}{cluster to use for parallel apply; see \link[parallel]{makeCluster}.}

\item{fit_negative}{logical; if \code{TRUE}, negative values are replace by NA.}

\item{RasterLayer}{logical; if \code{TRUE}, return a RasterLayer object.}

\item{...}{Passed on to \link[stars]{read_stars} parameters.}
}
\description{
A three-step algorithm for creating a mosaic from satellite imagery (see notes).
}
\section{Note}{

The three steps are: \cr
\itemize{
   \item Chooses the image less affected by atmospheric noise.
   \item Calculate the time gap between the image selected in step one to the remaining images.
   \item Take the clean pixels  considering the shortest time gap and ending with the image most temporally distant.
}
}

\section{References}{

Tarazona, Y., Mantas, V.M., Pereira, A.J.S.C. (2018). Improving tropical
deforestation detection through using photosynthetic vegetation time
series – (PVts-b). Ecological Indicators, 94, 367 379.
}

\examples{
library(ForesToolboxRS)
library(raster)
library(stars)
time <- as.Date(c("2016-07-30","2016-08-15","2016-09-16"))
rasterio = list(nXOff = 250, nYOff = 250, bands = c(4,3,2))

## Create a mosaic using three landsat images (just one band)
img <- system.file("simple_mosaic", package="ForesToolboxRS") \%>\%
  list.files("\\\\.tif$",full.names = TRUE)
free_img <- MosaicFreeCloud(img,time)
plot(free_img)

## Create a mosaic using three landsat images (all the bands)
img <- system.file("mosaic", package="ForesToolboxRS") \%>\%
  list.files("\\\\.tif$",full.names = TRUE)
free_img <- MosaicFreeCloud(img,time,RasterIO=rasterio)
plotRGB(free_img,3,2,1)

# Recommend way to use MosaicFreeCloud
\dontrun{
img <- system.file("mosaic", package="ForesToolboxRS") \%>\%
  list.files("\\\\.tif$",full.names = TRUE) \%>\%
  read_stars(img,RasterIO=rasterio) \%>\%
  merge
CLUSTER = parallel::makeCluster(spec = 4)
mosaic <- MosaicFreeCloud(img, time,CLUSTER = CLUSTER)
plotRGB(free_img,3,2,1)
}
}
