% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionRS.R
\name{print.fusionRS}
\alias{print.fusionRS}
\title{Print of the "fusionRS" class}
\usage{
\method{print}{fusionRS}(x, ...)
}
\arguments{
\item{x}{Object of class "fusionRS".}

\item{...}{Not used.}
}
\description{
Print of the "fusionRS" class
}
