% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pvts.R
\name{print.pvts}
\alias{print.pvts}
\title{Print of the "pvts" class}
\usage{
\method{print}{pvts}(x, ...)
}
\arguments{
\item{x}{Object of class "pvts".}

\item{...}{Not used}
}
\description{
Print of the "pvts" class
}
