% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvts.R
\name{pvts}
\alias{pvts}
\title{Change detection using the PVts-\eqn{\beta} approach}
\usage{
pvts(x, start, end, threshold = 5, img)
}
\arguments{
\item{x}{Matrix, RasterStack, RasterBrick}

\item{start}{The start of the monitoring time}

\item{end}{The end of the monitoring time}

\item{img}{The last monitoring image (in case "x" is a matrix)}

\item{threslhold}{The default threshold is 5 for photosynthetic vegetation,
while for indices such as NDVI and EVI the threshold is 3.
Please see Tarazona et al. (2018) for more details.}
}
\description{
This algorithm will allow to detect disturbances in the forests using
all the available Landsat set. In fact, it can also be run with sensors
such as MODIS.
}
\section{References}{

Tarazona, Y., Mantas, V.M., Pereira, A.J.S.C. (2018). Improving tropical
deforestation detection through using photosynthetic vegetation time
series – (PVts-β). Ecological Indicators, 94, 367 379.
}

\section{Note}{

In order to optimise the detections, it is advisable to make a smoothing
through the \link[ForesToolboxRS]{smootH} function before detecting changes. The smoothing will
allow to eliminate outliers that were not eliminated during the masking
of atmospheric artifacts.
}

\examples{
library(ForesToolboxRS)
library(raster)

## Not run:
imgs <- as.matrix(stack(list)) # imgs is a matrix
dim(imgs)[1]= 1000 is the number of pixels (time series)
dim(imgs)[2]=29 is the number of images between 1990 and 2018
Change monitoring period 2015-2018. 2015(position 26),2018(position 29)

cd <- pvts(x=imgs, start=26, end=29, threshold=5,img=lastimg)
plot(cd)

}
