% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkmeans.R
\name{rkmeans}
\alias{rkmeans}
\title{Unsupervised classification - kmeans}
\usage{
rkmeans(
  img,
  k,
  iter.max = 100,
  nstart = 50,
  algo = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{img}{RasterStack or RasterBrick}

\item{k}{the number of clusters}

\item{iter.max}{The maximum number of iterations allowed}

\item{nstart}{if centers is a number, how many random sets should be chosen?}

\item{algo}{It can be "Hartigan-Wong", "Lloyd", "Forgy" or "MacQueen". See \link[stats]{kmeans}}

\item{verbose}{This paramater is Logical. It Prints progress messages during execution.}

\item{...}{Options to be passed to the function. See \link[stats]{kmeans}}
}
\description{
This function allows to classify satellite images using k-means.
}
\details{
In principle, this function allows to classify satellite images specifying
a \code{k} value, however it is recommended to find the optimal value of \code{k} using
the \link[ForesToolboxRS]{calkmeans} function.

If warnings such as "Quick-TRANSfer stage steps exceeded maximum" or
"did not converge in 10 iterations" are obtained, it will be necessary to increase the
iterations in 20 or 30 (i.e., \code{inter.max = 20} or \code{iter.max = 30}). This issue is usually
obtained with "Hartigan-Wong". See details of \link[stats]{kmeans}.
}
\examples{
library(ForesToolboxRS)

# Load the dataset
data(img_l8)

# Select the best embedded algorithm in kmeans
classKmeans <- rkmeans(img = img_l8, k = 4, algo = "MacQueen")
}
\references{
Gareth James, Daniela Witten, Trevor Hastie, Robert Tibshirani. (2013).
An introduction to statistical learning : with applications in R. New York: Springer.
}
