% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{serie_pv}
\alias{serie_pv}
\title{A CLASlite example}
\format{
A RasterBrick object with 26 layers representing a time-series
}
\usage{
serie_pv
}
\description{
A dataset containing a photosynthetic vegetation time series obtained from CLASlite.
}
\keyword{datasets}
