% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smootH.R
\name{smootH}
\alias{smootH}
\title{Smoothing time series}
\usage{
smootH(x, Interp = "na.interp")
}
\arguments{
\item{x}{Numeric, matrix}

\item{Interp}{Four interpolation methods are presented, "na.interp",
"na.StructTS", "na.approx" and "na.spline". By default is the method
"na.interp".}
}
\description{
In order to eliminate outliers in the time series, Hamunyela et al. (2013)
smoothing is used (see References).
}
\section{References}{

Hamunyela, E., Verbesselt, J., Roerink, G., & Herold, M. (2013).
Trends in spring phenology of western European deciduous forests.
Remote Sensing,5(12), 6159-6179.

Tarazona, Y., Mantas, V.M., Pereira, A.J.S.C. (2018). Improving tropical
deforestation detection through using photosynthetic vegetation time
series – (PVts-β). Ecological Indicators, 94, 367 379.
}

\examples{
library(ForesToolboxRS)
library(forecast)
library(zoo)

x <- c(80,78,75,76,79,-100,82,76,81,77,76)
smth <- smootH(x)
plot(x, type="o", ylab="Reflectance \%", xlab="Time")
lines(smth, col="blue", type="o")

}
