% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForesToolboxRS-package.R
\docType{package}
\name{ForesToolboxRS-package}
\alias{ForesToolboxRS}
\alias{ForesToolboxRS-package}
\title{ForesToolboxRS: Remote Sensing Tools for Forest Monitoring}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Remote sensing tools for forest monitoring includes a series of 
    novel techniques to monitor forest through optical images and fusion of optical 
    and microwave data. All tools are based on scientific publications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ytarazona.github.io/ForesToolboxRS/}
  \item Report bugs at \url{https://github.com/ytarazona/ForesToolboxRS/issues}
}

}
\author{
\strong{Maintainer}: Yonatan Tarazona Coronel \email{geoyons@gmail.com} (\href{https://orcid.org/0000-0002-5208-1004}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

Other contributors:
\itemize{
  \item Alaitz Zabala \email{alaitz.zabala@uab.cat} (\href{https://orcid.org/0000-0002-3931-4221}{ORCID}) [contributor]
  \item Xavier Pons \email{xavier.pons@uab.cat} (\href{https://orcid.org/0000-0002-6924-1641}{ORCID}) [contributor]
  \item Antoni Broquetas \email{broquetas@tsc.upc.edu} (\href{https://orcid.org/0000-0001-9801-9145}{ORCID}) [contributor]
  \item Hamdi A. Zurqani \email{hzurqan@clemson.edu} (\href{https://orcid.org/0000-0002-9250-4534}{ORCID}) [contributor]
}

}
\keyword{internal}
