% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla.R
\name{mla}
\alias{mla}
\title{Supervised and unsupervised classification in Remote Sensing}
\usage{
mla(img, endm, model, training_split = 80, verbose = FALSE, ...)
}
\arguments{
\item{img}{RasterStack or RasterBrick}

\item{endm}{Signatures. SpatialPointsDataFrame or SpatialPolygonsDataFrame (typically shapefile)
containing the training data}

\item{model}{Model to use. It can be Support Vector Machine (\link[e1071]{svm}) like
\code{model = 'svm'}, Random Forest (\link[randomForest]{randomForest})
like \code{model = 'randomForest'}, Naive Bayes (\link[e1071]{naiveBayes})
like \code{model = 'naiveBayes'}, Decision Tree (\link[caret]{train})
like \code{model = 'LMT'}, Neural Networks (\link[nnet]{nnet})
like \code{model = 'nnet'}, K-nearest Neighbors (\link[caret]{knn3}) like \code{model = 'knn'}.}

\item{training_split}{For splitting samples into two subsets, i.e. training data and for
testing data}

<<<<<<< HEAD
\item{verbose}{This paramater is Logical. It Prints progress messages during execution.}

\item{...}{Parameters to be passed in machine learning algorithm. Please see \link[e1071]{svm}, \link[randomForest]{randomForest},
\link[e1071]{naiveBayes}, \link[caret]{train}, \link[nnet]{nnet} and \link[caret]{knn3}.}
=======
\item{verbose}{This parameter is Logical. It Prints progress messages during execution}
>>>>>>> be6d0b1 (cleans code styles and documentation)
}
\description{
It allows to execute supervised classification
in satellite images through various algorithms.
}
\section{Note}{
 If \code{model = "LMT"}, the function is using "Logistic Model Trees"
from http://topepo.github.io/caret/train-models-by-tag.html of the \code{caret} package.
}

\examples{
library(ForesToolboxRS)
library(raster)
library(snow)
library(caret)

# Load the dataset
data(FTdata)

# Random Forest Classifier
classMap <- mla(img = image, endm = endm, model = "randomForest", training_split = 80)
print(classMap)

# Plot image
plot(classMap$Classification)

}
\references{
Gareth James, Daniela Witten, Trevor Hastie, Robert Tibshirani. (2013).
An introduction to statistical learning : with applications in R. New York: Springer.

Mountrakis, G., Im, J., Ogole, C. (2011). Support vector machines in remote sensing:
A review. ISPRS Journal of Photogrammetry and Remote Sensing, 66, 247-259.

Belgiu, M., Dragut., L. (2016). Random Forest in Remote Sensing: A Review of Applications
and Future Directions. ISPRS Journal of Photogrammetry and Remote Sensing, 114, 24-31.

Maxwell, A.E., Warner, T.A., Fang, F. (2018). Implementation of machine-learning
classification in remote sensing: an applied review. International Journal of Remote
Sensing, 29(9), 2784-2817.

Pradhan, R., Ghose, M.K., Jeyaram, A. (2010). Land Cover Classification of Remotely
Sensed Satellite Data using Bayesian and Hybrid classifier. International Journal
of Computer Applications, 7(11).

Holloway, J., Mengersen, K. (2018). Statistical Machine Learning Methods and Remote
Sensing for Sustainable Development Goals: A Review. Remote Sensing, 10(9), 1365.
}
