% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calkmeans.R
\name{plot.calkmeans}
\alias{plot.calkmeans}
\title{Plot of the "calkmeans" class}
\usage{
\method{plot}{calkmeans}(x, xlab, ylab, type, main, cex, ...)
}
\arguments{
\item{x}{Object of class "calkmeans".}

\item{xlab}{X-axis title.}

\item{ylab}{Y-axis title.}

\item{type}{Type of line.}

\item{main}{Title of the graph.}

\item{cex}{Graph line point size.}

\item{...}{Graphical parameters to be passed in \link[base]{plot}.}
}
\description{
Plot of the "calkmeans" class
}
