% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pvts.R
\name{plot.pvts}
\alias{plot.pvts}
\title{Plot of the "pvts" class}
\usage{
\method{plot}{pvts}(x, ylab, type, col, lwd, main, ...)
}
\arguments{
\item{x}{Object of class "pvts".}

\item{ylab}{Y-axis title.}

<<<<<<< HEAD
\item{type}{Type of line.}

\item{col}{Type of color line.}

\item{lwd}{Line thickness size.}

\item{main}{Title of the graph.}

\item{...}{Graphical parameters to be passed in \link[base]{plot}.}

\item{cex}{Graph line point size.}
=======
\item{type}{Line type.}

\item{col}{Line color.}

\item{lwd}{Line thickness.}

\item{main}{Title.}

\item{...}{Additional graphical parameters to be passed to \link[base]{plot}.}
>>>>>>> 1b58dd1 (updates data)
}
\description{
Plot of the "pvts" class
}
