% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pvts.R
\name{plot.pvts}
\alias{plot.pvts}
\title{Plot for the "pvts" class}
\usage{
\method{plot}{pvts}(x, ylab, type, col, lwd, main, ...)
}
\description{
Plot for the "pvts" class
}
\author{
Yonatan Tarazona Coronel
}
