% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla.R
\name{print.mla}
\alias{print.mla}
\title{Print of the "mla" class}
\usage{
\method{print}{mla}(x, ...)
}
\arguments{
\item{x}{Object of class "mla".}

\item{...}{Not used.}
}
\description{
Print of the "mla" class
}
