% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvts.R
\name{pvts}
\alias{pvts}
\title{Change detection using the PVts-\eqn{\beta} approach}
\usage{
pvts(x, startm, endm, threshold = 5, verbose = FALSE)
}
\arguments{
\item{x}{Vector, univariate time series (class "ts") or a matrix without NA's.}

\item{startm}{The start of the monitoring time.}

\item{endm}{The end of the monitoring time.}

\item{threshold}{The default threshold is 5 for photosynthetic vegetation,
while for indices such as NDVI and EVI the threshold is 3.
Please see Tarazona et al. (2018) for more details.}

\item{verbose}{This paramater is Logical. It Prints progress messages during execution.}
}
\description{
This algorithm will allow to detect disturbances in the forests using
all the available Landsat set. In fact, it can also be run with sensors
such as MODIS.
}
\section{References}{

Tarazona, Y., Mantas, V.M., Pereira, A.J.S.C. (2018). Improving tropical
deforestation detection through using photosynthetic vegetation time
series (PVts-\eqn{\beta}). Ecological Indicators, 94, 367 379.
}

\section{Note}{

In order to optimise the detections, it is advisable to make a smoothing
through the \link[ForesToolboxRS]{smootH} function before detecting changes. The smoothing will
allow to eliminate outliers that were not eliminated during the masking. If the object is a
time series, it must have the structure of a "ts", that is, the serie must have a start, end
and frequency. See \link[stats]{ts} for more details. In addition, in case the input is a
matrix, the first dimension must be rows*columns of the image, and the second dimension the
number of images.
}

\examples{
library(ForesToolboxRS)

# Example 1.
# Normalized Difference Fraction Index from 1990 to 2017, one index for each year.
ndfi <- c(0.86,0.93,0.97,0.91,0.95,0.96,0.91,0.88,0.92,0.89,
         0.90,0.89,0.91,0.92,0.89,0.90,0.92,0.84,0.46,0.20,
         0.27,0.22,0.52,0.63,0.61,0.67,0.64,0.86)

# Detect changes in 2008 (position 19)
cd <- pvts(x = ndfi, startm = 19, endm = 19, threshold = 5)
plot(cd)

# Detect changes in 2008 (time serie)
vec <- ts(ndfi, start = 1990, end = 2017, frequency = 1)
cd <- pvts(x = ndfi, startm = 2008, endm = 2008,  threshold = 5)
plot(cd)

}
\author{
Yonatan Tarazona Coronel
}
