% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkmeans.R
\name{rkmeans}
\alias{rkmeans}
\title{Unsupervised classification - kmeans}
\usage{
rkmeans(
  img,
  k,
  iter.max = 100,
  nstart = 50,
  algo = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  ...
)
}
\arguments{
\item{img}{RasterStack or RasterBrick}

\item{k}{the number of clusters}

\item{iter.max}{The maximum number of iterations allowed}

\item{nstart}{if centers is a number, how many random sets should be chosen?}

\item{algo}{It can be "Hartigan-Wong", "Lloyd", "Forgy" or "MacQueen". See \link[stats]{kmeans}}

\item{...}{Options to be passed to the function. See \link[stats]{kmeans}}

\item{verbose}{This paramater is Logical. It Prints progress messages during execution}
}
\description{
This function allows to classify satellite images using k-means.
}
\details{
If we want to find the optimal value of \code{k} (clusters or classes),
we should use \code{k = NULL}. This approach will look for the k values for
which the intra-class inertia is stabilized. If we know the \code{k} value and the idea
is to find the best algorithm embedded in k-means, that maximizes inter-class distances,
we must put \code{k = n}, where \code{n} is a specified number of classes.
}
\examples{
library(ForesToolboxRS)

# Load the dataset
data(FTdata)

# Selecting the best embedded algorithm in kmeans
classKmeans <- rkmeans(img = image, k = 4, algo = "MacQueen")
}
\references{
Gareth James, Daniela Witten, Trevor Hastie, Robert Tibshirani. (2013).
An introduction to statistical learning : with applications in R. New York: Springer.
}
