% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.R
\name{sma}
\alias{sma}
\title{Spectral Mixture Analysis (SMA)}
\usage{
sma(img, endm)
}
\arguments{
\item{img}{Optical images. It could be RasterStack or RasterBrick.}

\item{endm}{Endmembers must be a matrix or data.frame and with more than one edmember.
Rows represent the endmembers and columns represent the spectral bands.  The number
of bands must be greater than the number of endmembers.}
}
\description{
The SMA assumes that the energy received, within the field of vision of the
remote sensor, can be considered as the sum of the energies received from each
dominant endmember. This function addresses a Linear Mixing Model.
}
\section{References}{

Adams, J. B., Smith, M. O., & Gillespie, A. R. (1993). Imaging spectroscopy:
Interpretation based on spectral mixture analysis. In C. M. Pieters & P.
Englert (Eds.), Remote geochemical analysis: Elements and mineralogical
composition. NY: Cambridge Univ. Press 145-166 pp.

Shimabukuro, Y.E. and Smith, J., (1991). The least squares mixing models to
generate fraction images derived from remote sensing multispectral data.
IEEE Transactions on Geoscience and Remote Sensing, 29, pp. 16-21.
}

\section{Note}{

A regression analysis is used to obtain the fractions. In least squares
inversion algorithms, the common objective is to estimate abundances that
minimize the squared error between the actual spectrum and the estimated spectrum.
The values of the fractions will be between 0 and 1.
}

\examples{
\dontrun{
library(ForesToolboxRS)
library(raster)
library(dplyr)

#Load an example dataset
data(fdata)

#Endmembers
soil<-c(0.8980,0.8508,0.8704,1.3636,1.6579,1.1420)
forest<-c(0.8207,0.7545,0.6548,1.6463,0.9725,0.6673)
water<-c(0.9276,0.9570,1.0089,0.6743,0.5220,0.5143)
endm <- matrix(c(soil,forest,water), 3, 6, byrow = TRUE, dimnames =
list(c("soil", "forest","water"), c("B1", "B2", "B3","B4","B5","B6")))

#Unmix the image
fractions <- sma(img_ndfi, endm)
}
}
