% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smootH.R
\name{smootH}
\alias{smootH}
\title{Smoothing time series}
\usage{
smootH(x, method.interp = "na.interp")
}
\arguments{
\item{x}{Numeric, matrix.}

\item{method.interp}{Four interpolation methods are presented, "na.interp",
"na.approx" and "na.spline". By default is the method "na.interp".}
}
\description{
In order to eliminate outliers in the time series, a temporary smoothing is used.
}
\details{
Vegetation indices or fractions of photosynthetic activity generally present
noise in the time series that in some way or another hinder subsequent analyzes.
This noise can be due to two reasons: i) no algorithm was used to mask the atmospheric
noise before analyzing the time series and ii) negative outliers that were not detected
by the masking algorithm. To minimize this problem, it is possible to eliminate negative
outliers using the method proposed by [Hamunyela et al. (2013)](https://doi.org/10.3390/rs5126159).

This method is not able to remove consecutive outliers. The mathematical approach
of this method of removing outliers implies the non-modification of the first and last
values of the historical series, so that the near real-time detections of ecosystem
disturbances will not be affected in any way.
}
\section{References}{

Tarazona, Y., Mantas, V.M., Pereira, A.J.S.C. (2018). Improving tropical
deforestation detection through using photosynthetic vegetation time
series (PVts-\eqn{\beta}). Ecological Indicators, 94, 367 379.

Hamunyela, E., Verbesselt, J., Roerink, G., & Herold, M. (2013).
Trends in spring phenology of western European deciduous forests.
Remote Sensing,5(12), 6159-6179.
}

\examples{

library(ForesToolboxRS)

# Here a vector with an outlier
x <- c(80,78,75,76,79,-100,82,76,81,77,76)
plot(x, type="o", col="red")

# Applying a smoothing
smth <- smootH(x)
plot(x, type="o", ylab="Reflectance \%", xlab="Time")
lines(smth, col="blue", type="o")

}
\author{
Yonatan Tarazona Coronel
}
